chore: check bb.issueComments.create and bb.issueComments.update
