+++ frontend/src/components/IssueV1/components/ActivitySection/ActivityList.vue
@@ -66,7 +66,7 @@
       </ActivityItem>
     </ul>
 
-    <div v-if="!state.editCommentMode">
+    <div v-if="!state.editCommentMode && allowCreateComment">
       <div class="flex">
         <div class="flex-shrink-0">
           <div class="relative">
@@ -120,9 +120,14 @@ import {
 } from "@/types";
 import type { ComposedIssue } from "@/types";
 import { LogEntity, LogEntity_Action } from "@/types/proto/v1/logging_service";
-import { extractUserResourceName } from "@/utils";
+import { extractUserResourceName, hasProjectPermissionV2 } from "@/utils";
 import { doSubscribeIssue, useIssueContext } from "../../logic";
-import { ActivityItem, DistinctActivity, isSimilarActivity } from "./Activity";
+import {
+  ActivityItem,
+  DistinctActivity,
+  isSimilarActivity,
+  isUserEditableActivity,
+} from "./Activity";
 
 interface LocalState {
   editCommentMode: boolean;
@@ -203,6 +208,14 @@ const activityList = computed((): DistinctActivity[] => {
   return distinctActivityList;
 });
 
+const allowCreateComment = computed(() => {
+  return hasProjectPermissionV2(
+    issue.value.projectEntity,
+    currentUser.value,
+    "bb.issueComments.create"
+  );
+});
+
 const cancelEditComment = () => {
   state.activeActivity = undefined;
   state.editCommentMode = false;
@@ -228,19 +241,17 @@ const doCreateComment = async (comment: string) => {
 };
 
 const allowEditActivity = (activity: LogEntity) => {
-  if (activity.action !== LogEntity_Action.ACTION_ISSUE_COMMENT_CREATE) {
+  if (!isUserEditableActivity(activity)) {
     return false;
   }
-  if (currentUser.value.email !== extractUserResourceName(activity.creator)) {
-    return false;
+  if (currentUser.value.email === extractUserResourceName(activity.creator)) {
+    return true;
   }
-  const payload = JSON.parse(
-    activity.payload
-  ) as ActivityIssueCommentCreatePayload;
-  if (payload && payload.externalApprovalEvent) {
-    return false;
-  }
-  return true;
+  return hasProjectPermissionV2(
+    issue.value.projectEntity,
+    currentUser.value,
+    "bb.issueComments.update"
+  );
 };
 
 const onUpdateComment = (activity: LogEntity) => {
